Source packaging improvements.
Write a MANIFEST.in so that `python setup.py sdist` will ship the
documentation and licensing files and test.py.  MANIFEST is now a
generated file, so remove it from git.

(This should fix the other problem reported in issue #11, README.rst
being missing when setup.py tries to read it.)