Read README.rst as UTF-8 regardless of external locale settings.
The sanitized environment used by Debian package build daemons forces
use of the barebones ‘POSIX’ locale.  Some versions of Python 3 will
refuse to read an UTF-8 file as text when this locale is in use,
unless explicitly told to use UTF-8 anyway.  Python 2 doesn’t support
the encoding= argument to open(), so to get setup.py to do the Right
Thing in a universal fashion, read the file as *binary* and then
explicitly decode it.