Change package version number from 0.8.3 to 0.996.1.
There are two reasons for this change.  Most important, the SWIG
definitions we ship are copied out of the MeCab library’s source tree
(see <https://github.com/taku910/mecab/tree/master/mecab/swig>).
This makes clear which version of those definitions we have.

Also, Debian already ships a version of these bindings that are
numbered mostly consistently with the MeCab library (python-mecab,
currently version 0.99.6-2) but are stuck without Python 3 support.
I would like to try to replace them with this package, but if we
stick to version 0.8.3, they will have to put an “epoch” on their
package numbering in order to do that, and they prefer to avoid that
whenever possible.