Add the ability to download and build a local copy of libmecab.
This is needed in order to produce binary wheels that do not depend on
a system-provided libmecab (as required by PEP 513 for Linux, and
practical constraints for OSX and Windows).  The new mode is activated
by setting BUNDLE_LIBMECAB=1 in the environment when running setup.py.

(Note: it will still be necessary to install a dictionary by hand.)