Configure tox and pytest; revise travis.yml to use tox.
In order to bundle a dictionary, we are going to need to be able to
detect its presence or absence from MeCab/__init__.py, and this cannot
work correctly when running out of the development tree.  Set up tox
so we can easily test the installed package instead.  This also gets
us some basic validation of the source distribution.