Switch to modern SWIG codegen for Python.
SWIG’s default code generation for Python accommodates CPython
versions all the way back to 2.0, which we do not need.  `-O -builtin`
code generation works fine with 2.7 and later, has less runtime
overhead, and eliminates the need for the machine-generated Python
wrapper file (`MeCab.py`).  This last will make it easier to institute
some hand-written shims that tell libmecab where to find its
dictionary, when the dictionary is embedded in the wheel.

One unfortunate side effect is that one of the C++-level `Tagger::parse`
overloads has to be excluded from wrapper generation, because the
modern code generation cannot handle overload sets in which some but
not all methods are `static` (in the C++ sense).