When libmecab is to be bundled, also bundle a dictionary.
The dictionary must already exist, in compiled form, on the build
system.  setup.py can be pointed at it with any of several environment
variables; failing that it will look for a mecabrc in the usual
locations and read the dicdir setting from that file.

At run time, MeCab/__init__.py detects that a dictionary has been
bundled into the package, and directs the library to use it.  This can
be overridden by an existing setting of MECABRC in os.environ, or by
arguments to the Tagger and Model constructors.