.travis.yml: only run deployment if the build succeeded.
Otherwise we can get a partial upload from a build that fell over
after building some but not all of the wheels, which prevents us
from just triggering a rebuild if the problem was a transient glitch
(e.g. Travis build workers occasionally fail to download stuff that we
know exists).