Fix sdist generation
The sdist generation code was written before support for manylinux1 was
added. As a result the build attempts to upload the sdist twice, which
causes an error. This change makes it so only the manylinux2010 build
uploads the sdist.

The choice of which Linux build to use is basically arbitrary, but since
manylinux2010 is likely to be around longer it seemed like a reasonable
choice.