Add correct argument handling
Old argument handling couldn't handle quoted arguments. This handles
quoted arguments in Python and passes an array of strings to the C++
MeCab code.

Most of the complexity in this change is not in Python but in SWIG. Most
details are handled by a SWIG typedef converter.

This should fix #25 and #43.