Add tip about trying args with Model
See #45 for details. For some reason the Tagger constructor doesn't
produce error strings, but the Model constructor does. Since the arg
format is the same you can get an error by passing it to the Model.

May want to automate that if it turns out this is an upstream bug that
can't be fixed.